pyproject.toml
[project]
name = "instagram-image-splitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "inquirer>=3.4.0",
    "pillow>=11.1.0",
]
